Speed up job existence tests by fetching less info
Most methods are asserting that a job exists by calling get_job_info().
Unfortunately that might trigger on the server side a parse of the full
job history to gather the lastSuccessful, lastFailure build numbers.
Since Jenkins is heavily file based, that the process is rather long on
jobs having thousands and thousands of build.

Since we only care about checking whether the job exists, use
job_exists() which is light on the server side.  That speed up
operations tremendously whenever the targetted jobs are very long.

Change-Id: I946dbb7a08a8bb172a369af3b2aedc07e3cea9e6