Add a get_plugin_info method
This change adds a method to get jenkins plugin information. It may be called
without any parameters which will return a list of all installed plugins on
master. User may specify the plugin name in which case it will return info
about that specific plugin. If the plugin name cannot be found then it will
return None.

Change-Id: Ia6cd90571ff260aec9869f48df06486b7fac181f