basic support for folder plugin
Support traversing folders in jenkins to retrieve job info, determine if
jobs exist based on the last component of the name if it's a job in a
folder and return the full name of the job instead of the basename.

Current includes support for amending the information returned by the
jenkins server to include the fullname of the job '<item>/<name>' so
that essentially any container of jobs can be referenced uniquely
through it's fullname attribute.

Can perform create and deletion, and includes a fix for quoting of
spaces in the names of jobs.

Change-Id: I8b27f4d1d1ddc9a55ada8569282d8c56502286a2