Raise an exception on 404
Currently, when Jenkins returns a 404 error, the jenkins_open()
method gives no indication.  This adds a new NotFoundException,
descending from JenkinsException, which can be use to distinguish
this case.  This causes a problem with cancel_queue(), however,
as the Jenkins API always returns a 404; since cancel_queue() is
a best-effort mechanism, this change also modifies that method to
explicitly ignore a NotFoundException.

Change-Id: I98320d174659fc83d48614b31a5a054fd0773dfe
Closes-Bug: #1246468