Eliminate redundant calls to assert_job_exists()
Now that the jenkins_open() method will raise a specific exception
if Jenkins returns a 404, some calls to assert_job_exists() now
become redundant.  Since these calls represent an extra round-trip
with the Jenkins server, dropping them will result in a reduction
in time taken for such things as a jenkins-job run.  This change
eliminates the redundant calls and adjusts the tests as appropriate.

Change-Id: If78dea9ad60d446854d24919bdd7f3f1fee3d733