Ensure request data is in the correct format
Ensure that the parameters passed to POST requests are in the correct format.
As with change Ia6ce75958bb, this change ensures that the parameters passed to
POST requests are in the correct format so the data needed to trigger the use
of POST won't result in an exception being thrown due to use of str instead
of unicode on python 3.

Change-Id: I542d56b6c392a8390d9eea6a860c798eade3f925