Fix tests on python 2.6 and mark as deprecated
Pin mock to < 1.1 as newer versions no longer work on python 2.6.
Remove duplicate unittest2 for python 2.6 test env.
Adding warning to mark python 2.6 support as deprecated and intended to
be removed.

Testtools re-added the requirement for unittest2 so the separate
dependency for python 2.6 created a duplicate error that causes tox
to fail.

Change-Id: Ia79eeeb5d2432cf1025631a1722de5fb1babe2c9