Add function to get running builds
It can be useful to get a list of running builds from jenkins. This is
particularly useful when you need to take an action on a specific built
or if you need to know when the jenkins master has finished running all
builds (say after quieting down). Add a function to return this list.

Change-Id: I5c7dac8076250f94cd2b358c5b153d9410f53aee