Use str as base class for PluginVersion
To allow PluginVersion class to be passed to builtin functions expecting
a string type, need to use 'str' as the base class. This ensures
the change to support version comparison is completely transparent to
the calling code, but permits changes to benefit from the new behaviour
without breaking existing usage.

Change-Id: I7d434294fa45171e47ea4876d01398cb75f5d962