Add optional kerberos support
Check if "kerberos" package can be imported at runtime (in __init__.py).
If not, everything works as before. If "kerberos" package can be
imported register a new urllib handler (defined in urllib_kerb.py) which
adds support for kerberos auth.

This urllib handler (urllib_kerb.py) is only triggered on 401 (Unauthorized)
response, and we try to auth with kerberos. If unsuccessful, next 401 handler
(if any) is triggered.

Change-Id: I1a47e455aa14535a124df950994718a11d7e4f57