better get_build_info() example
The "nextBuildNumber" attribute from get_build_info() does not always
predict the result of the next build_job() invocation. If the Jenkins
server has many jobs queued or in flight, the job from the next
build_job() call may end up running as an even higher job number.

A better way to track the result of a build_job() is to poll the job's
queue item and wait for it to return the exact job number.

Change-Id: I830f9c83cc9940d477f03758f4764f3209f0ccce