Adds support for multi-select build parameters
This feature will allow build_job to accept build parameters as a list of two membered tuples - [('parameter key', 'parameter value', ), ...].
Currently, build_job expects to receive a dictionary of parameters but this limits the build_job_url function (which is called from build_job()) to build jenkins url for jobs with multi-select parameter(s) because dictionaries cannot have duplicate keys (in case, when there are several values to one parameter key). urlencode is capable of converting lists with two membered tuples to url as well.

Change-Id: I134a5f8b69ba91acd89afa791cb1cf0b5e63f5f5