Enables use of PYTHONHTTPSVERIFY override
Since we switched to requests library, SSL verification
changed. In order to avoid breaking users that were
using non globally recognized certificates we
detect if PYTHONHTTPSVERIFY=0 is defined and disable
requests SSL verfication.

This has no effect if this variable is not defined as
we do not want to have any insecure defaults.

Change-Id: I2ea7ebb97f806e915b5b96a3f9ec47d48b039ad2