enable cross testing with jenkins-job-builder
Adds a new job that tests if change affects jenkins-job-builder
master branch in order to avoid introducing breaking changes
in the most important consumer of the this library.

Also adds an optional tox target named tips which can be used
by developers to perform the same kind of testing. This
mimics the same kind of testing that was already implemented
inside jenkins-job-builder.

Change-Id: I594347c431a881227c5090b41b4ba4ef50e3133f
Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>